move imageitem to outer class
