+++ app/src/main/java/im/bsk/storyalbum/GalleryActivity.java
@@ -24,6 +24,7 @@ import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import im.bsk.storyalbum.adapters.ImagePagerAdapter;
 import permissions.dispatcher.NeedsPermission;
 import permissions.dispatcher.OnNeverAskAgain;
 import permissions.dispatcher.OnPermissionDenied;
