+++ app/src/main/java/im/bsk/storyalbum/NewStoryActivity.java
@@ -10,22 +10,18 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.common.collect.Collections2;
 import com.mikepenz.fastadapter.adapters.FastItemAdapter;
-import com.mikepenz.fastadapter.items.AbstractItem;
-import com.mikepenz.fastadapter.utils.ViewHolderFactory;
-import com.squareup.picasso.Picasso;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import im.bsk.storyalbum.items.ImageItem;
 
 public class NewStoryActivity extends AppCompatActivity {
 
@@ -147,59 +143,4 @@ public class NewStoryActivity extends AppCompatActivity {
         }
     }
 
-    static class ImageItem extends AbstractItem<ImageItem, ImageItem.ViewHolder> {
-        private static final ViewHolderFactory<? extends ViewHolder> FACTORY = new ItemFactory();
-
-        public final File file;
-
-        public ImageItem(File file) {
-            this.file = file;
-        }
-
-        //The unique ID for this type of item
-        @Override
-        public int getType() {
-            return R.id.fastadapter_new_story_image_id;
-        }
-
-        //The layout to be used for this type of item
-        @Override
-        public int getLayoutRes() {
-            return R.layout.item_new_story_image;
-        }
-
-        @Override
-        public void bindView(final ViewHolder viewHolder, List payloads) {
-            super.bindView(viewHolder, payloads);
-
-            Picasso.with(viewHolder.itemView.getContext())
-                    .load(file)
-                    .placeholder(R.mipmap.ic_launcher)
-                    .fit()
-//                    .centerCrop()
-                    .into(viewHolder.image);
-
-        }
-
-        protected static class ItemFactory implements ViewHolderFactory<ViewHolder> {
-            public ViewHolder create(View v) {
-                return new ViewHolder(v);
-            }
-        }
-
-        @Override
-        public ViewHolderFactory<? extends ViewHolder> getFactory() {
-            return FACTORY;
-        }
-
-        protected static class ViewHolder extends RecyclerView.ViewHolder {
-            @BindView(R.id.item_new_image)
-            ImageView image;
-
-            public ViewHolder(View view) {
-                super(view);
-                ButterKnife.bind(this, itemView);
-            }
-        }
-    }
 }
